+++ lib/src/date_adapter.dart
@@ -3,7 +3,7 @@ import 'abstract_date.dart';
 abstract class DateAdapter<T extends AbstractDate> {
   T fromDateTime(DateTime dateTime);
 
-  DateTime toDateTime(T date);
+  DateTime toDateTime(AbstractDate date);
 
   int monthLength(int year, int month);
 }