+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.0.3]
+- added default adapters `HijriDate`,
+ `GregorianDate` or `MiladiDate` ,
+  `ShamsiDate` or `JalaliDate` or `PersianDate`
+
 ## [0.0.2]
 - added `isTypeOf<T>()` to `Date`